Create PlaceManagerTestCase
Tests timing out
